Advanced search view user results added
