+++ mini_githubcic/mini_githubcic/templates/advanced_search_result.html
@@ -4,7 +4,13 @@
     <div class="col-auto mr-auto" style="display:flex;">
         <div style="border-bottom: groove">
             {% block advanced_search %}
+            {% if search_type == 's3' %}
             {% include "advanced_search.html" with s1=True s2=True s3=True project_id=project_id user_id=user_id %}
+            {% elif search_type == 's2' %}
+            {% include "advanced_search.html" with s1=True s2=True s3=False project_id=project_id user_id=user_id %}
+            {% else %}
+            {% include "advanced_search.html" with s1=True s2=False s3=False project_id=project_id user_id=user_id %}
+            {% endif %}
             {% endblock advanced_search %}
         </div>
         <div style="margin-top: 8px;margin-left:20px">Search results for "{{keyword}}" in
@@ -66,12 +72,12 @@
         <div class="item1">
             {% if selected is 1 %}
             {% include "advanced_search_result_projects.html" with projects=projects %}
-            {% endif %}
-            {% if selected is 2 %}
+            {% elif selected is 2 %}
             {% include "advanced_search_result_issues.html" with issues=issues %}
-            {% endif %}
-            {% if selected is 3 %}
+            {% elif selected is 3 %}
             {% include "advanced_search_result_pr.html" with prs=prs %}
+            {% else %}
+            {% include "advanced_search_result_users.html" with users=users %}
             {% endif %}
         </div>
     </div>
