+++ mini_githubcic/mini_githubcic/templates/advanced_search_result_users.html
@@ -1,0 +1,30 @@
+{% if users %}
+<div class="row" style="border-bottom-style: inset;width: 100%;margin-left: 1px;">
+    <div class="col-auto mr-auto" >
+        <h3> {{ users|length }} user results</h3>
+    </div>
+</div>
+<br>
+{% for user in users %}
+<div class="card mb-3">
+    <div class="card-body row">
+        <div class="col-auto mr-auto">
+            <p class="card-text">{{ user }} </p>
+            <a class="btn btn-primary" href="{% url 'profile_preview' user.username %}">View</a>
+        </div>
+    </div>
+</div>
+{% endfor %}
+{% else %}
+<div data-view-component="true"
+     style="padding: 32px;position: relative;text-align: center;padding-right: 8px !important;padding-left: 8px) !important;">
+    <svg aria-hidden="true" height="24" viewBox="0 0 24 24" version="1.1" width="24" data-view-component="true"
+         class="octicon octicon-search blankslate-icon">
+        <path fill-rule="evenodd"
+              d="M10.25 2a8.25 8.25 0 105.28 14.59l5.69 5.69a.75.75 0 101.06-1.06l-5.69-5.69A8.25 8.25 0 0010.25 2zM3.5 10.25a6.75 6.75 0 1113.5 0 6.75 6.75 0 01-13.5 0z"></path>
+    </svg>
+    <h3 data-view-component="true" class="mb-1">We could not perform this search</h3>
+    <p>None of the search qualifiers apply to this search type.</p>
+
+</div>
+{% endif %}
